changing recurring order + some refactoring
