+++ assets/js/src/components/admin/ShortcodesDocs.js
