+++ assets/js/src/components/public/EventDetails.js
